+++ src/interfaces/Options.ts
@@ -1,8 +1,6 @@
-import { Info } from '../types/Info';
-
-export interface Options<T> {
-  name: string;
-  version?: string;
-  parseOutput?: boolean;
-  info?: T[] | Info[];
+export interface Options {
+  readonly name: string;
+  readonly version?: string;
+  readonly parseOutput?: boolean;
+  readonly info?: string[] | string;
 }