+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # npm-get-package-info Changelog
 
+## 2.0.0 (2020-05-19)
+#### Bug Fix
+- fixed type for custom package.json fields
+- fixed output for multiple package versions
+- fixed type when none of values is valid info item
+- fixed different output when only one info item is passed
+
 ## 1.0.0 (2020-04-19)
 #### New Feature
 - added `npmGetPackageInfo` with `name`, `version`, `parseOutput` & `info` options